Termino do ajuste do sistema das pastas e os makefiles para o mesmo utilizado pelo rasa boilerplate.
