+++ .gitbook.yaml
@@ -1,0 +1,2 @@
+structure:
+  readme: README.md
