--- Rasa/actions/actions.py
@@ -1,13 +1,0 @@
-
-from rasa_core_sdk import Action
-
-
-class ActionTest(Action):
-    def name(self):
-        return "action_test"
-
-    def run(self, dispatcher, tracker, domain):
-        try:
-            dispatcher.utter_message("Mensagem enviada por uma custom action.")
-        except ValueError:
-            dispatcher.utter_message(ValueError)
