--- Rasa/coach/policy_config.yml
@@ -1,13 +1,0 @@
-policies:
-  - name: "KerasPolicy"
-    featurizer:
-    - name: MaxHistoryTrackerFeaturizer
-      max_history: 5
-      state_featurizer:
-        - name: BinarySingleStateFeaturizer
-  - name: "MemoizationPolicy"
-    max_history: 5
-  - name: "FallbackPolicy"
-    nlu_threshold: 0.4
-    core_threshold: 0.3
-    fallback_action_name: "utter_default"
