--- Rasa/data/intents/chitchat.md
@@ -1,286 +1,0 @@
-## intent:cumprimentar
-- com licenca
-- Com licença
-- licença
-- Licença
-- posso falar com voce
-- Posso falar com voce
-- posso falar com você
-- Posso falar com você
-- pode me tirar uma duvida
-- Pode me tirar uma duvida
-- pode me tirar uma dúvida
-- Pode me tirar uma dúvida
-- gostaria de tirar uma duvida
-- gostaria de tirar uma dúvida
-- Gostaria de tirar uma duvida
-- Gostaria de tirar uma dúvida
-- bom dia
-- Bom dia
-- boa tarde
-- Boa tarde
-- boa noite
-- Boa noite
-- bonjour
-- bonsoir
-- olá
-- Olá
-- ola
-- Ola
-- oi
-- Oi
-- oiii
-- Oiii
-- oiiii
-- Oiiii
-- oiiiii
-- Oiiiii
-- oiiiiii
-- Oiiiiii
-- oie
-- Oie
-- oiee
-- Oiee
-- ola boa tarde
-- Ola boa tarde
-- hi
-- Hi
-- Fala
-- fala
-
-## intent:despedir
-- tchau obrigada
-- Tchau obrigada
-- tchau obrigado
-- Tchau obrigado
-- tchau
-- Tchau
-- namaste
-- sayonara
-- au revoir
-- até logo
-- Até logo
-- ate logo
-- Ate logo
-- até mais
-- Até mais
-- ate mais
-- Ate mais
-- até breve
-- Até breve
-- ate breve
-- Ate breve
-- até a próxima
-- Até a proxima
-- ate a proxima
-- Ate a proxima
-- falou, valeu
-- Falou, valeu
-- flw vlw
-- Flw vlw
-- beleza
-- Beleza
-- blza
-- Blza
-- blz
-- Blz
-- valew
-- Valew
-- valeu
-- Valeu
-- vlw
-- Vlw
-- flw
-- Flw
-
-## intent:tudo_bem
-- como vai voce
-- como vao as coisas
-- opa tudo bem
-- tudo bem
-- tudo bom
-- como voce esta
-- como voce ta
-- como vai você
-- como vai voce
-- como vai vc
-- como vai
-- como ta
-
-## intent:preocupacao
-- oi, tudo bem
-- oi tudo bem
-- oiiii, tudo bem
-- oiiii tudo bem
-- Ola tudo bem
-- ola tudo bem
-- Ola, tudo bem
-- ola, tudo bem
-- eae blz
-- Eae blz
-- eae beleza
-- Eae beleza
-- eae, beleza
-- Eae, beleza
-- Eae, blz
-- eae, blz
-
-## intent:explicacao
-- como funciona
-- Como funciona
-- e como funciona
-- O que é você
-- não entendi
-- nao entendi
-
-## intent:segredo
-- é uma cilada
-- É uma cilada
-- Eh uma cilada
-- eh uma cilada
-- eh uma silada
-- Eh uma silada
-- É uma silada
-- é uma silada
-- É uma cilada Bino
-- é uma silada Bino
-- É uma silada Bino
-- é uma silada Bino
-- Eh uma cilada bino
-- eh uma cilada bino
-- É uma cilada bino
-- é uma silada bino
-- É uma silada bino
-- é uma silada bino
-- Eh uma cilada bino
-- eh uma cilada bino
-
-## intent:risada
-- piada
-- engraçado
-- kkk
-- kkkk
-- kkkkk
-- kkkkkk
-- Hahahaha
-- Haha
-- hahaha
-- hahahaha
-- hahahahaha
-
-## intent:animes1
-- Omae Wa Mou Shindeiru
-- omae wa mou shindeiru
-
-## intent:animes2
-- Anime
-- anime
-- mangá
-- Mangá
-
-## intent:xingamento
-- Me Mama
-- me mama
-- mama
-- Mama
-- Me mama
-- me Mama
-- vai se fuder
-- fodace
-- quero que você se foda
-- caguei
-- se fuder
-- foda-se
-- vai tomar no cu
-- toma no cu
-- cu
-- filho da puta
-- filha da puta
-- fila da puta
-- ti fuder
-- te fuder
-- Te fuder
-- Ti fuder
-- Desgraçado
-- Seu desgraçado
-- desgraçado
-- seu desgraçado
-
-## intent:reacao_feliz
-- sim
-- Sim
-- Estou bem
-- estou bem
-- bem
-- Bem
-- Bom
-- bom
-- Bonzin
-- bonzin
-- tranquilo
-- Tranquilo
-- Estou otimo
-- estou otimo
-- Tudo certo
-- tudo certo
-- Tudo tranquilo
-- tudo tranquilo
-- Dboa
-- dboa
-- de boa
-- De boa
-- Sussa
-- sussa
-- tranks
-- Tranks
-- Suave
-- suave
-- joia
-- joinha
-
-## intent:reacao_triste
-- Nao
-- nao
-- estou pessimo
-- Estou pessimo
-- Nao estou bem
-- Nao estou bem
-- Queria estar melhor
-- queria estar melhor
-- estou mal
-- Estou mal
-- estou triste
-- Estou triste
-- mais ou menos
-- Mais ou menos
-- cansado
-- Cansado
-- Cansada
-- cansada
-- infeliz
-- Infeliz
-- ja tive dias melhores
-- Ja tive dias melhores
-- muito triste
-- Muito triste
-- desanimado
-- Desanimado
-- desanimada
-- Desanimada
-
-## intent:series
-- serie
-- Serie
-- seriado
-- Seriado
-- series
-- Series
-- gosto de series
-- assistir series
-- ver series
-
-## intent:futebol
-- futebol
-- Futebol
-- gosta futebol
-- esporte
