--- Rasa/data/intents/commands.md
@@ -1,15 +1,0 @@
-## intent:help
-- help
-- /help
-
-## intent:atividades
-- atividades
-- /atividades
-
-## intent:provas
-- provas
-- /provas
-
-## intent:trabalhos
-- trabalhos
-- /trabalhos
