--- Rasa/endpoints.yml
@@ -1,16 +1,0 @@
-action_endpoint:
-  url: "http://actions:5055/webhook"
-
-# tracker_store:
-#     type: mongod
-#     url: mongodb://mongo:27017/
-#     db: telegram_db
-#     username: admin
-#     password: "1234"
-
-
-tracker_store:
-  type: mongod
-  url: mongodb://mongo:27017
-  # username: rasa
-  # password: example
