+++ bot/credentials.yml
@@ -1,0 +1,16 @@
+#telegram:
+#  access_token: ${TELEGRAM_TOKEN}
+#  verify: ${TELEGRAM_BOT_USERNAME}
+#  webhook_url: ${TELEGRAM_WEBHOOK}
+
+#rocketchat:
+#  user: ${ROCKETCHAT_BOT_USERNAME}
+#  password: ${ROCKETCHAT_BOT_PASSWORD}
+#  server_url: ${ROCKETCHAT_URL}
+
+socketio:
+  user_message_evt: user_uttered
+  bot_message_evt: bot_uttered
+  session_persistence: true
+  
+rest:
