--- constantes.py
@@ -1,2 +1,0 @@
-TELEGRAM_TOKEN = "962521399:AAGyRWTL9kAmjcgFn6mem_DxDyeXcbMRppA"
-TELEGRAM_DB = "mongodb://mongo:27017/telegram_db"
