--- docker/bot-telegram.env
@@ -1,3 +1,0 @@
-TELEGRAM_BOT_USERNAME=binoUnB_bot
-TELEGRAM_TOKEN=962521399:AAGyRWTL9kAmjcgFn6mem_DxDyeXcbMRppA
-TELEGRAM_WEBHOOK=https://d99d413e.ngrok.io/webhooks/telegram/webhook
