+++ docker/bot.Dockerfile
@@ -1,11 +1,7 @@
-FROM lappis/coach:boilerplate as coach
-FROM lappis/botrequirements:boilerplate
-
-
-COPY ./Rasa/bot /bot
-COPY --from=coach /src_models/ /models/
+FROM botrequirements
 
 WORKDIR /bot
 
+COPY ./bot /bot
 
 RUN find . | grep -E "(__pycache__|\.pyc|\.pyo$)" | xargs rm -rf
