--- docker/train.Dockerfile
@@ -1,5 +1,0 @@
-FROM rasa/rasa:latest-full
-
-COPY ./Rasa .
-
-RUN train --domain domain.yml --data data --out models --config config.yml
