+++ requirements.txt
@@ -1,4 +1,5 @@
 rasa==1.4.2
 nltk==3.4.5
+elasticsearch==7.0.4
 flake8==3.7.8
 pymongo==3.9.0
